+++ src/gpfa_controller.f90
@@ -1,4 +1,5 @@
 module gpfa_controller
+   use, intrinsic :: iso_fortran_env, only: real64
    use gpfa_dispatch, only: gpfa
    implicit none
 contains
@@ -7,10 +8,10 @@ contains
 !      use iso_c_binding
       implicit none
       integer :: m, isign, nblk, n, i, inc
-      real(8) :: trig(*), cr(nblk * m), ci(nblk * m)
-!      real(8), pointer :: cr(:)
-!      real(8) :: cr(2*nblk*m)
-!      complex(8), target :: c(nblk*m)
+      real(real64) :: trig(*), cr(nblk * m), ci(nblk * m)
+!      real(real64), pointer :: cr(:)
+!      real(real64) :: cr(2*nblk*m)
+!      complex(real64), target :: c(nblk*m)
 !      call C_F_POINTER(C_LOC(c), cr, [2*nblk*m])
 !      cr(1:2*nblk*m-1:2)=dble(c(1:nblk*m))
 !      cr(2:2*nblk*m:2)=aimag(c(1:nblk*m))
@@ -21,7 +22,7 @@ contains
          i = n
          call gpfa(cr(i:), ci(i:), trig, inc, 1, m, 1, isign)
       end do
-!      c(1:nblk*m)=cmplx(cr(1:2*nblk*m-1:2),cr(2:2*nblk*m:2),kind=kind(0.0d0))
+!      c(1:nblk*m)=cmplx(cr(1:2*nblk*m-1:2),cr(2:2*nblk*m:2),kind=real64)
    end subroutine cgpfa
 
 end module gpfa_controller
